Removed unused package call with the custom toolchain
As @cachebag pointed out, I was defining the method call with a custom
toolchain attribute set but didn't use it when calling the `buildPackage` function.

Removed the `naersk-lib` definition, and renamed `naersk'` to
`naersk-package` for a clearer name and used it to build the package
later.

Whether or not this actually builds the package as needed requires more
understanding on my part, but it should build a binary just fine for all
targets that `nix` can run on.